AssetBuilder and rework some parts of the renderer

+++ src/Titan.Graphics/Buffers/IIndexBuffer.cs
@@ -5,6 +5,6 @@ namespace Titan.Graphics.Buffers
     public interface IIndexBuffer : IResource, IDisposable
     {
         public uint NumberOfIndices { get; }
-        public ref readonly short[] Indicies { get; }
+        public ref readonly int[] Indicies { get; }
     }
 }
